chore: update various class names
