+++ Tests/RudderStackAnalyticsTests/Helpers/MockProvider.swift
@@ -7,7 +7,6 @@
 
 import Testing
 import Foundation
-import Network
 @testable import RudderStackAnalytics
 
 // MARK: - MockProvider
@@ -146,189 +145,6 @@ final class MockProvider {
     }
 }
 
-// MARK: - Unified Mock Plugin for Swift Testing
-final class MockEventCapturePlugin: Plugin {
-    var pluginType: PluginType
-    weak var analytics: Analytics?
-    
-    // Tracking properties
-    private(set) var setupCalled = false
-    private(set) var executeCalled = false
-    private(set) var shutdownCalled = false
-    private(set) var capturedEvents: [Event] = []
-    private let eventLock = NSLock()
-    
-    // Configuration properties for advanced testing
-    var shouldFilterEvent = false
-    var shouldModifyEvent = false
-    var eventModifications: [String: Any] = [:]
-    
-    // MARK: - Initialization
-    
-    init(type: PluginType = .terminal, enableFiltering: Bool = false, enableModification: Bool = false) {
-        self.pluginType = type
-        self.shouldFilterEvent = enableFiltering
-        self.shouldModifyEvent = enableModification
-    }
-    
-    // MARK: - Plugin Protocol Methods
-    
-    func setup(analytics: Analytics) {
-        self.analytics = analytics
-        setupCalled = true
-    }
-    
-    func intercept(event: any Event) -> (any Event)? {
-        executeCalled = true
-        
-        // Always capture the event (before any filtering/modification)
-        eventLock.lock()
-        capturedEvents.append(event)
-        eventLock.unlock()
-        
-        // Handle filtering
-        if shouldFilterEvent {
-            return nil
-        }
-        
-        // Handle event modification
-        if shouldModifyEvent {
-            var modifiedEvent = event
-            
-            // Apply modifications based on event type
-            if var trackEvent = modifiedEvent as? TrackEvent {
-                var properties = trackEvent.properties?.dictionary?.rawDictionary ?? [:]
-                for (key, value) in eventModifications {
-                    properties[key] = value
-                }
-                trackEvent.properties = CodableCollection(dictionary: properties)
-                modifiedEvent = trackEvent
-            } else if var screenEvent = modifiedEvent as? ScreenEvent {
-                var properties = screenEvent.properties?.dictionary?.rawDictionary ?? [:]
-                for (key, value) in eventModifications {
-                    properties[key] = value
-                }
-                screenEvent.properties = CodableCollection(dictionary: properties)
-                modifiedEvent = screenEvent
-            } else if var identifyEvent = modifiedEvent as? IdentifyEvent {
-                var traits = identifyEvent.traits?.dictionary?.rawDictionary ?? [:]
-                for (key, value) in eventModifications {
-                    traits[key] = value
-                }
-                identifyEvent.traits = CodableCollection(dictionary: traits)
-                modifiedEvent = identifyEvent
-            } else if var groupEvent = modifiedEvent as? GroupEvent {
-                var traits = groupEvent.traits?.dictionary?.rawDictionary ?? [:]
-                for (key, value) in eventModifications {
-                    traits[key] = value
-                }
-                groupEvent.traits = CodableCollection(dictionary: traits)
-                modifiedEvent = groupEvent
-            }
-            
-            return modifiedEvent
-        }
-        
-        return event
-    }
-    
-    func shutdown() {
-        shutdownCalled = true
-    }
-    
-    // MARK: - Event Access Methods
-    
-    var lastProcessedEvent: Event? {
-        eventLock.lock()
-        defer { eventLock.unlock() }
-        return capturedEvents.last
-    }
-    
-    var receivedEvents: [Event] {
-        eventLock.lock()
-        defer { eventLock.unlock() }
-        return capturedEvents
-    }
-    
-    func getEventsOfType<T: Event>(_ type: T.Type) -> [T] {
-        eventLock.lock()
-        defer { eventLock.unlock() }
-        return capturedEvents.compactMap { $0 as? T }
-    }
-    
-    func clearEvents() {
-        eventLock.lock()
-        capturedEvents.removeAll()
-        eventLock.unlock()
-    }
-    
-    var eventCount: Int {
-        eventLock.lock()
-        defer { eventLock.unlock() }
-        return capturedEvents.count
-    }
-    
-    // MARK: - Configuration Methods
-    
-    func enableFiltering() {
-        shouldFilterEvent = true
-    }
-    
-    func disableFiltering() {
-        shouldFilterEvent = false
-    }
-    
-    func enableModification(with modifications: [String: Any] = [:]) {
-        shouldModifyEvent = true
-        eventModifications = modifications
-    }
-    
-    func disableModification() {
-        shouldModifyEvent = false
-        eventModifications = [:]
-    }
-    
-    func setEventModifications(_ modifications: [String: Any]) {
-        eventModifications = modifications
-    }
-}
-
-extension MockEventCapturePlugin {
-    // Generic version (wait for specific type)
-    @discardableResult
-    func waitForEvents<T: Event>(_ type: T.Type, count expectedCount: Int = 1, timeout: TimeInterval? = nil) async -> [T] {
-        let start = Date()
-        
-        while true {
-            let events = getEventsOfType(type)
-            if events.count >= expectedCount {
-                return events
-            }
-            if let timeout, Date().timeIntervalSince(start) > timeout {
-                return events
-            }
-            await Task.yield()
-        }
-    }
-
-    // Non-generic version (wait for all events)
-    @discardableResult
-    func waitForEvents(count expectedCount: Int = 1, timeout: TimeInterval? = nil) async -> [Event] {
-        let start = Date()
-        
-        while true {
-            let events = receivedEvents
-            if events.count >= expectedCount {
-                return events
-            }
-            if let timeout, Date().timeIntervalSince(start) > timeout {
-                return events
-            }
-            await Task.yield()
-        }
-    }
-}
-
 // MARK: - Test Helpers
 extension MockProvider {
     
@@ -344,15 +160,6 @@ extension MockProvider {
         URLProtocol.unregisterClass(MockURLProtocol.self)
         let config = URLSessionConfiguration.ephemeral
         HttpNetwork.session = URLSession(configuration: config)
-    }
-    
-    func runAfter(_ seconds: Double, block: @escaping () async -> Void) async {
-        // Suspend the current task for the specified duration
-        let nanoseconds = UInt64(seconds * 1_000_000_000)
-        try? await Task.sleep(nanoseconds: nanoseconds)
-        
-        // Execute the block after the delay
-        await block()
     }
     
     static var sourceConfiguration: SourceConfig? {
@@ -381,59 +188,6 @@ extension MockProvider {
         let config = URLSessionConfiguration.ephemeral
         config.protocolClasses = [MockURLProtocol.self]
         return URLSession(configuration: config)
-    }
-    
-    static func prepareMockDataPlaneSession(with responseCode: Int) -> URLSession {
-        MockURLProtocol.requestHandler = { _ in
-            let json = responseCode == 200 ? ["Success": "Ok"] : ["error": "Server error"]
-            let data = try JSONSerialization.data(withJSONObject: json)
-            return (responseCode, data, ["Content-Type": "application/json"])
-        }
-        
-        let config = URLSessionConfiguration.ephemeral
-        config.protocolClasses = [MockURLProtocol.self]
-        return URLSession(configuration: config)
-    }
-}
-
-// MARK: - MockLogger for capturing log output
-class SwiftMockLogger: Logger {
-    var logs: [(level: String, message: String)] = []
-    
-    func verbose(log: String) {
-        logs.append(("VERBOSE", log))
-    }
-    
-    func debug(log: String) {
-        logs.append(("DEBUG", log))
-    }
-    
-    func info(log: String) {
-        logs.append(("INFO", log))
-    }
-    
-    func warn(log: String) {
-        logs.append(("WARN", log))
-    }
-    
-    func error(log: String, error: Error?) {
-        if let error {
-            logs.append(("ERROR", "\(log) - \(error.localizedDescription)"))
-        } else {
-            logs.append(("ERROR", log))
-        }
-    }
-    
-    func clearLogs() {
-        logs.removeAll()
-    }
-    
-    func hasLog(level: String, containing message: String) -> Bool {
-        return logs.contains { $0.level == level && $0.message.contains(message) }
-    }
-    
-    func logCount(for level: String) -> Int {
-        return logs.filter { $0.level == level }.count
     }
 }
 
@@ -464,61 +218,6 @@ extension MockProvider {
         event.anonymousId = "<anonymous-id>"
         event.originalTimestamp = "<original-timestamp>"
     }
-}
-
-// MARK: - MockAnalytics
-class MockAnalytics: Analytics {
-    var isFlushed: Bool = false
-    
-    init() {
-        let config = Configuration(writeKey: "_sample_write_key_", dataPlaneUrl: "_sample_data_plane_url_")
-        super.init(configuration: config)
-    }
-    
-    override func flush() {
-        self.isFlushed = true
-    }
-}
-
-// MARK: - MockNetworkMonitor
-class MockNetworkMonitor: NetworkMonitorProtocol {
-    var status: NWPath.Status
-    var interfaces: [NWInterface.InterfaceType]
-    
-    init(status: NWPath.Status, interfaces: [NWInterface.InterfaceType]) {
-        self.status = status
-        self.interfaces = interfaces
-    }
-    
-    func usesInterfaceType(_ type: NWInterface.InterfaceType) -> Bool {
-        return interfaces.contains(type)
-    }
-    
-    func start(queue: DispatchQueue) {
-        // Simulate path update
-    }
-    
-    func cancel() {
-        // Simulate cancel behavior
-    }
-}
-
-// MARK: - String(Extension)
-extension String {
-    var trimmed: String {
-        return self.trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-}
-
-// MARK: - Run After
-
-func runAfter(_ seconds: Double, block: @escaping () async -> Void) async {
-    // Suspend the current task for the specified duration
-    let nanoseconds = UInt64(seconds * 1_000_000_000)
-    try? await Task.sleep(nanoseconds: nanoseconds)
-    
-    // Execute the block after the delay
-    await block()
 }
 
 // MARK: - MockProvider(Extension)
@@ -555,4 +254,22 @@ extension MockProvider {
         "Firebase": true,
         "Braze": false
     ]
+}
+
+// MARK: - String(Extension)
+extension String {
+    var trimmed: String {
+        return self.trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+}
+
+// MARK: - Run After
+
+func runAfter(_ seconds: Double, block: @escaping () async -> Void) async {
+    // Suspend the current task for the specified duration
+    let nanoseconds = UInt64(seconds * 1_000_000_000)
+    try? await Task.sleep(nanoseconds: nanoseconds)
+    
+    // Execute the block after the delay
+    await block()
 }
